add dto
