+++ src/components/card/card.tsx
@@ -1,45 +1,57 @@
-import type { CardProps } from '../../types';
+type Card = {
+  src: string;
+  premium: boolean;
+  bookmarks: boolean;
+  price: number;
+  rating: number;
+  description: string;
+  type: string;
+}
 
-export default function Card({ src, premium, bookmarks, price, rating, description, type }: CardProps): JSX.Element {
+type CardProps = {
+  card: Card;
+}
+
+export default function Card({ card }: CardProps): JSX.Element {
 
   return (
     <article className="cities__card place-card">
 
-      { premium && (
+      { card.premium && (
         <div className="place-card__mark">
           <span>Premium</span>
         </div>) }
 
       <div className="cities__image-wrapper place-card__image-wrapper">
         <a href="#">
-          <img className="place-card__image" src={src} width="260" height="200" alt="Place image" />
+          <img className="place-card__image" src={card.src} width="260" height="200" alt="Place image" />
         </a>
       </div>
       <div className="place-card__info">
         <div className="place-card__price-wrapper">
           <div className="place-card__price">
-            <b className="place-card__price-value">&euro;{price}</b>
+            <b className="place-card__price-value">&euro;{card.price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
           <button className="place-card__bookmark-button button" type="button">
             <svg className="place-card__bookmark-icon" width="18" height="19">
               <use xlinkHref="#icon-bookmark"></use>
             </svg>
-            {bookmarks && (
+            {card.bookmarks && (
               <span className="visually-hidden">To bookmarks</span>
             )}
           </button>
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
-            <span style={{ width: `${rating}%` }}></span>
+            <span style={{ width: `${card.rating}%` }}></span>
             <span className="visually-hidden">Rating</span>
           </div>
         </div>
         <h2 className="place-card__name">
-          <a href="#">{description}</a>
+          <a href="#">{card.description}</a>
         </h2>
-        <p className="place-card__type">{type}</p>
+        <p className="place-card__type">{card.type}</p>
       </div>
     </article>
   );
