+++ src/components/cards-list/cards-list.tsx
@@ -22,7 +22,7 @@ export default function CardsList({ data }: { data: PlaceType[] }): JSX.Element 
     <section className="cities__places places">
       <h2 className="visually-hidden">Places</h2>
       <b className="places__found">312 places to stay in Amsterdam</b>
-      < Sort />
+      <Sort />
       <div className="cities__places-list places__list tabs__content">
         {offers}
       </div>
